+++ src/components/ProjectCard.js
@@ -40,7 +40,7 @@ import EditCharacterProject from './EditCharacterProject';
 import EditItemProject from './EditItemProject';
 import ItemButton from './ItemButton';
 import data_utils from '../scripts/data_utils';
-import { MODAL_STYLE } from '../scripts/constants';
+import _CONSTANTS, { MODAL_STYLE } from '../scripts/constants';
 
 /**
  * MANAGES PROJECT CARDS AND THEIR DESIGN.
@@ -94,11 +94,14 @@ export default function ProjectCard({ data, project, defaultEnabled, projectMapD
     useEffect(() => {
         // this will be updated via key change if inventory changed
         const invCheck = data_utils(data).project.check(project, userState.inventory,
-            userState.settings.use_legacy ? 0 : undefined,
+            userState.settings.use_legacy
+                ? _CONSTANTS.RECIPE_NOTE.LEGACY
+                : userState.settings.use_legacy_2
+                    ? _CONSTANTS.RECIPE_NOTE.LEGACY_2 : undefined,
             project.details.ignored_rarity || {});
         setCompleted(invCheck[0]);
         updateRequiredAndMissing({ project, invCheck });
-    }, [project, userState.settings.use_legacy, userState.inventory, data]);
+    }, [project, userState.settings.use_legacy, userState.settings.use_legacy_2, userState.inventory, data]);
 
     // hide/show project details
     function handleExpand() {